third_party/blink/web_tests/external/wpt/css/css-highlight-api/highlight-pseudo-computed-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle() for ::highlight(foo): should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo)) should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo)( should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo)(foo) should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo)() should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for :::highlight(foo) should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo). should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo,bar) should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
[FAIL] getComputedStyle() for ::highlight(foo bar) should not return a style.
  assert_equals: Invalid pseudo identifiers should not return a style. expected 0 but got 390
Harness: the test ran to completion.
